Remove a pd2l simplification handler in J9
Remove the replaceSetSignUnderBinaryWithAbs() handler for the following
two reasons:

1. This handler handles obsolete pd2iu and pd2lu opcodes. The Java DAA API
will never generate these unsigned nodes.

2. This handler can also introduce labs (long absolute value) nodes above
pd2l nodes. This is problematic in two ways: first, this handler only
replaces the pd2l with labs on the first appearance of the pd2l. Subsequent
references to that pd2l are not fixed up. If an pd2l is replaced by labs,
the current pd2l and all future references to it must all be replaced.
Second, the BCDCHK evaluator, by design, is not able
to recognize nor handle none PD related nodes. As a result of this, having
labs nodes under BCDCHK nodes can cause runtime crashes.

Signed-off-by: Nigel Yu <yunigel@ca.ibm.com>