Remove java.version check in getProperty tests for jdk9+
The java.version system property is set by OpenJDK code so it can be
assumed to be correct. The hard coded value in these tests would need to
be updated every 6 months for each Java version as well.
Therefore, the java.version check will be removed from the getProperty
tests.
Also changes the order of java version conditionals in build.xml so
future version of Java use the Java 9 compile path with modules.

Signed-off-by: Mike Zhang <mike.h.zhang@ibm.com>