Prevent register spilling inside internal control flow
We were not preventing the register assigner from assigning
R0 in the evaluator that inlines Atomic operations like
AtomicInteger.getAndDecrement(). Some virtual registers
are used in addi and stwcx which can't use R0 forcing
the register assigner to spill inside the internal control
flow generated by the evaluator if R0 was assigned.

Signed-off-by: Kevin Langman <langman@ca.ibm.com>