Improve GC pause times by removing assumptions in one pass
Use a mark and sweep approach to removing runtime assumptions
from the Runtime Assumption Table (RAT) in places where we are
doing assumption clean up during a GC pause time. This will
significantly reduce the GC pause times in cases were we have
a large number of assumptions and the GC is unloading classes.

Signed-off-by: Kevin Langman <langman@ca.ibm.com>