Fix a crash in findClassInfo on PPC
Due to PPCs weak memory coherency it's possible for a crash
to occur in findClassInfo() if the caller has not obtained
the class table lock. This fix changes call to findClassInfo
into calls to findClassInfoAfterLocking in places where
the class table lock is not held. Also fixed a compiler
warning in j9method.cpp.

Signed-off-by: Kevin Langman <langman@ca.ibm.com>