Do not archive test material in pr builds
Because we were archiving in parallel to testing,
archiving the test material would often fail in PR
builds becasue we are trying to write to the test
folder at the same time we are trying to tar it.
This PR will change the behaviour so we do not
bother archiving the test material in PR builds. Since we
compile & test on the same machine in PRs. In non-pr
builds we do compile on one machine and then test on
another.

[skip ci]

Fixes #1114

Signed-off-by: Adam Brousseau <adam.brousseau88@gmail.com>