Fixed Inline Array Allocation
When inlining array allocations, several checks are performed to make sure that
the allocation is valid and can be done inline. Some of the checks performed
try to determine if the size or number of elements of the array that is being
allocated fall under a max value. If they do not, the VM handles the allocation
out-of-line.

An issue with the aformentioned checks was that the size or number of elements,
were being stored in registers that could have most significant bits that were
contaminated by  prior instructions. Thus when the checks were performed,
although the size of the array or number of elements was within the range of an
inline array allocation, it would seem as if they were not and an out-of-line
allocation would occur, affecting the performance.

In order to fix this, the bits that could be contaminated are cleared as to not
affect any following calculations.

Signed-off-by: Ryan Santhirarajan <rsanth@ca.ibm.com>