Rename ArgumentMoverHandle file to BruteArgumentMoverHandle
Rename the file so the class referenced by other classes is the name of
the file.

As part of the refactoring for JDK10, replace all referenes to
ArgumentMoverHandle with BruteArgumentMoverHandle so only one of the two
classes is referenced outside the file.

Move several methods from Argument to Brute MoverHandle.  These are static
helper methods that are referenced from other classes.

Signed-off-by: Dan Heidinga <daniel_heidinga@ca.ibm.com>