Implement X86 codegen support for dynamic array clone
Dynamic array cloning transforms refArrayObject.clone()
into `anewarray` followed by `arraycopy`, where refArrayObject
is a resolved, not fixed class. For code gen, It requires the
class (2nd) child of anewarray to be runtime evaluated instead
of compile time constant, as actual refArrayObject type is not
known at compiled time (other than it's an ref array). This
change implements support in X86 codegen to handle a runtime
evaluated class child of anewarray.

This is prerequisite for the optimizer change that performs the
object.clone() call transformation.

Signed-off-by: Yan Luo <Yan_Luo@ca.ibm.com>